Add WIn32 Logger. Better than log4cpp!

+++ Include/MyServer/Helper.h
@@ -104,7 +104,7 @@ public:
 	}
 
 	static std::string base64_decode(std::string const& encoded_string) {
-		int in_len = encoded_string.size();
+		size_t in_len = encoded_string.size();
 		int i = 0;
 		int j = 0;
 		int in_ = 0;
